Split components into two files
